+++ src/app/profile/page.tsx
@@ -1,4 +1,4 @@
-import ProfilePageClient from '@/components/profile/profile-page'
+import ProfilePageClient from '@/components/profile/profile-page-client'
 
 export default function ProfilePage() {
   return <ProfilePageClient />
